Fix race condition caused by unnecessary preview refreshes
We are calling `requestRefresh(...)` and, consequently, starting our
refresh coroutine flows when opening files containing Compose Previews.
The problem is we were doing that even if the project was not built.
This is problematic for a few reasons:

* We were starting the refresh flow as if we were about to render, but
  since that was not the case, our coroutine flow would be put in an
  inconsistent state, since callbacks would be executed ahead of time,
  assumptions would be broken, and so on.
* We were wasting unnecessary resources, because the computations for
  rendering the previews would still be made, even though nothing would
  render in the end.
* We'd show the "Refreshing Previews..." progress indicator  while the
  project still had to be compiled, resulting in a confusing UX.
* Class loading would throw a ClassNotFoundException when trying to load
  the framework classes, because  we'd try to render the preview before
  layoutlib being ready to render them.

This CL fixes the issue by preventing to call requestRefresh(...) until
the project is ready for that. Now, ComposePreviewRepresentationTest can
be re-enabled since it shouldn't time out because of this race condition
anymore. Other compose-designer tests might run faster, since we are now
expected to refresh fewer times when the project is not ready or has not
build yet.

Bug: 161091273
Test: ComposePreviewRepresentationTest and compose-designer tests
Change-Id: I3f2a6abbd537e254c03abc531f2ed9643f8310fb

GitOrigin-RevId: d2aa873ef9be16dd0bbf53a77222ee73a464bf36