Extend surround with widget action
This extends the detection behaviour and allows to surround widgets
without having a selection.
The change will infer a selection based on the caret position and will
use that as selection input for the surround actions.

The change also updates how the selection is used and will extend it to
the nearest call expression, allowing the selection to start in the
middle of, for example, the parameter list.

Fixes: 217696184
Test: ComposeSurroundWithWidgetActionTest
Change-Id: Ie746cb883d22c20f296544449434c9947d7a53a3

GitOrigin-RevId: b1c6ae55e8339abb424aeba6aaa885caf1cff790