Use main modules in Android run configurations
This is to allow things that try to get information from the module in
the run configuration to have a none empty module. This is
also the more "correct" setup as we are building from the sources of
this main module.

We also remove patching of run configurations from pre-qualified name
versions since with this we can no longer guess the module of the
runconfiguration. This was also broken in general after MPSS since we
run over all module based configuration whichs will set any none android
configuration to the empty holder module.

We also fix a small issue with Android test run configurations being
wrongly marked as invalid by adjusting the accepted modules for these
configurations.

Fixes: 217847811
Test: snapshots
Change-Id: Ib68d3f5e9b0a6740f111789900d41448c6804302

GitOrigin-RevId: f24e587a3f1981354adf66033b46efb0ce901107