Fix flake in testNonStructuralModelChanges
Before this change it was possible that the TreeModelListener added
towards the end of this test would be called as a result of an
earlier line in the test - because the earlier line resulted in
the callbacks being made on a different thread (the UI thread).

The fix is to wait and dispatch all UI events before adding the
callback.

Fixes: 217380465
Test: LayoutInspectorTreePanelTest.testNonStructuralModelChanges
Change-Id: I59491d55331538b029dc645c4cfa37486717af49

GitOrigin-RevId: dd1c80ca1225d90771bb6d39fb64f604765b9874