Updates to faster preview
Add "additional fast preview" actions drop down next to the full refresh
button. When fast preview is active, this drop down will allow disabling
the fast preview.

This change also updates the behaviour of fast preview so now, it will
refresh on change instead of on save.

Moved a number of checks of the fast preview flag to check the isEnabled
value in the LiveEditPreviewManager. This allows for these behaviours to
be disabled at runtime.

ProjectBuildStatusManager will now have a compiling state even if the
build was triggered by the fast preview.

Bug: 199384558
Test: ProjectBuildStatusManagerTest
Change-Id: Ia4818bb55bb668e97b83c3860f3dd310b5f92770

GitOrigin-RevId: 94b248d83a111678d7face210ef1c691c1c6df05