Updates to faster preview
Add "additional fast preview" actions drop down next to the full refresh
button. When fast preview is active, this drop down will allow disabling
the fast preview.

This change also updates the behaviour of fast preview so now, it will
refresh on change instead of on save.

Moved a number of checks of the fast preview flag to check the isEnabled
value in the LiveEditPreviewManager. This allows for these behaviours to
be disabled at runtime.

ProjectBuildStatusManager will now have a compiling state even if the
build was triggered by the fast preview.

Bug: 199384558
Test: ProjectBuildStatusManagerTest
Change-Id: I7aa20d6f3b9b1330f3dda01f6b0265bafbcfcf47

GitOrigin-RevId: 43e8c03c39cbb9859acf299e0d958677022b6a5d