Fix empty device list at startup
The initialization of DeviceExplorerController depended on the initial
device list being populated by the DebugBridgeChangeListener before
refreshDeviceList executed. This used to happen because the call to
refreshDeviceList was always enqueued in an EdtExecutor, giving the
DebugBridgeChangeListener a chance to run in between start and
refreshDeviceList; with coroutines, execution continues unless there is
a need to suspend.

Bug: 205005362
Test: updated
Change-Id: I2f2f83eb91d429b8f4d05e7275c954faf5389f0e

GitOrigin-RevId: 87fce51c2e2dd7620733a55c53f458f4ec1979be