Work around MDNS Openscreen adb issue
Per b/217251994, ADB crashes when ADB_MDNS_OPENSCREEN is set on
certain Windows configs. We attempt to work around this issue by
disabling ADB_MDNS_OPENSCREEN and notifying the user that ADB
WiFi is disabled for the next attempt at getting the debug bridge.

We restrict the work around to be very specific in order to avoid
any false positives.

Bug: 217251994
Test: manually via hardcoding an exception in startup connection
Change-Id: I4d388c95232996b59fd0daec95aee4c98fbe4cb1

GitOrigin-RevId: 4d09eb64633266c29e8040038d237feb9d6615dd