Refactor Preview toolbar action wrappers
Create a new wrapper class (RefreshingAndRenderErrorsActionWrapper) and
a helper function (disabledIfRefreshingOrRenderErrors()) that converts
all the actions in a list to use the wrapper. Use the helper function on
the list passed when creating the DefaultActionGroup that comprises the
Compose Preview toolbar, so all the actions in the group are properly
wrapped. We also change the call to visibleOnlyInComposeStaticPreview(),
so it's done in the entire list as opposed to each action individually.

Bug: 203426144
Test: N/A. Current tests should pass.
Change-Id: I1137a223a7068b43202d71d624a825b5e9a66c3a

GitOrigin-RevId: 23a0eec805a2626e72a90f7064b7be30d69a3932