Change Compose Preview look-up to first group by Composable method
The existing implementation of the process for finding the Compose
Previews of a file consists of first searching Preview annotations
using a kotlin index and then filtering out the methods that are
not also annotated with @Composable, which is inefficient in some
cases, making the time complexity O(#annotations-per-method ^ 2).

This change modifies the process of finding Compose Previews to
first search for Composable annotations using a kotlin index and
then process the Preview annotations of each of these Composable
methods. As a consequence, the time complexity will now be linear.

This new way of finding Previews is also needed by Multipreview,
an upcoming feature.

Bug: N/A
Test: existing tests pass and manually verified with a simple
compose project

Change-Id: I152449c622c7f40aff9181f24105b52dbc7659d3

GitOrigin-RevId: b60842628335737b6c64703fa21a551abce3f391