Make Gradle plugin resolution more dynamic in tests
Rather than checking the versions of plugins (other than
com.android.*) in the build files on creation to generate the plugins
resolution strategy for tests, resolve using the requested version.
This allows for tests to introduce a dependency on, for example, the
Kotlin plugin where none was existing initially; without this change
or one like it, attempts to resolve a newly-added plugin would resolve
to a dependency with version `null`.

Bug: 180411529
Bug: 217805224
Test: existing
Change-Id: I4d1daaeec38957902057fa73dbba72337ea0d17a

GitOrigin-RevId: 520035858349bec8c7796404891c4387f52b4362