Change ClearLogcatTask To Use an Interface Instead of User Data
Rather than have to Tool Window Factory inject a user data object, use an interface to identify the desired component.

This makes it possible for Logcat V2 to also use this task. There are a few issues with using user data in Logcat V2:
1. The user data is added to Content which in the Logcat V2 is not readily available since it is managed by the Splitting Tabs library.
2. ClearLogcatTask can't depend on Logcat V2 classes because Logcat V2 already depends on the core module (circular dependency).

As part of this cl, ClearLogcatTask is converted to Kotlin because it results in much cleaner code.

Bug: 191880898
Test: N/A
Change-Id: I437a326f0d3628eda44d3f1badb93f8c5dd93765

GitOrigin-RevId: b1dfeaf478d2672331a75306eb49ebf1ef557394