Fixes visual lint issue in shared issue panel
Change the display name of NoFileNode from "Others" to
"Layout Validation".
Fixes a missing function, which highlights the preview in Layout
Validation Tool when selecting an issue.

Also change the selection model of tree to single item, so user cannot
select multiple issues to highlight more previews in Layout Validation
Tool. (This may be useful, but we haven't defined this feature.)

Bug: 222110455
Test: covered and manually(ui-related)
Change-Id: I8ec8c061b13058a47eea765b32d5a723c0cc3830

GitOrigin-RevId: d43042211e9cf49d3757729a934cb90d1436e63e