Use backed file to save the zoom level of design surface
The file of NlModel may not be the same as editor's file. For example,
compose editor creates a temp layout file for rendering purpose.

The zoom level is saved by the file path and name. But the name or the
path of temp layout file may be changed time by time. We lose the
information of zoom level when new temp file is used.

Using the path of original file to save the zoom level to get rid of
this issue.

Fixes: 217382996
Test: add the unit test and test manually
Change-Id: I05cf2ef6bb1f0c6d9c24b54ca3ef25b8a6e44937

GitOrigin-RevId: cc3d1bc019f77281181e85e87fb33ff9dab62b21