Rename viewGroupHandler to layoutHandler
The extension method NlComponent.viewGroupHandler sounds like it
would return a ViewGroupHandler for the specified NlComponent.
But in reality it returns a handler for its parent layout unless
the component is a layout itself.
The extension method is renamed to layoutHandler to emphasize that
this is a handler for a layout that may be a parent.

In a followup change a viewGroupHandler is added with a different
meaning.

Bug: N/A
Test: N/A
Change-Id: I861fe32c51d18b0e862fee40f6f9e524b4e799f5

GitOrigin-RevId: f4c91ccf4fd3e27da67afb90772a2a954c1807cc