Remove SdkMerger.java
This also fixes an issue where the progress bar shown when installing
the SDK only goes up to 70%. This was because 30% was reserved for the
merge operation, which can't happen.

This is the second attempt at submitting this. The first coincided with
the failure of some flaky tests and was reverted.

Change-Id: Ia4627f0660416ae8de3f992779a1195473040970
Fixes: 222740445

GitOrigin-RevId: 0fe69cc8d7339f7de6a83f8b92b00388b8cfe952