Restrict parallel model fetching
We do not want to request parallel models on V1, and want to only
request V2's in parallel when using AGP >= 7.3-alpha04.
Moreover, we know thet non-android models fetching is not yet thread
safe, so in this change we are making sure that non-android models are
requested sequentially only.

Bug: 179156667
Bug: 220318679
Bug: 220325253
Bug: 215344823
Test: manually on Android Studio
Change-Id: Ia8a295cbda2d8e922a57fd9b45f793c47acd9bdd

GitOrigin-RevId: 0bed59006ef33182b09f5d9c94d0d8fea6f8fa28