[Dsl] Reintroduce HIDDEN element state
A previous version of the Dsl model used a HIDDEN state to
reflect (limited) semantics of operators, particularly those such as
.reset() to clear a collection.  Now, semantics of the model are
implemented, to the extent that they are implemented at all, by
walking the tree structures in the model corresponding to the Dsl
operators, so the HIDDEN state was unnecessary.

However, we now need the HIDDEN state to reflect transformations of
the tree structures in the model, in cases where there is not a
one-to-one correspondence of the structures themselves.  This is not
needed for the Dsl files themselves, but version catalog files
introduce two aspects which do require tree structure transformation:
alias normalization, and version references.

This change partially reverses change
I60adf530b6dd23554797dc891c6b97437108cab1 but should have no visible
effect at present.

Bug: 200280395
Test: existing (no change in functionality)
Change-Id: I7ebe97e470f8fb56e74f4cd4f7ff4340e19c5061

GitOrigin-RevId: 4fc3afa9e6e0a8fe5624bab229aebc6f811c3161