141578652: Remove Unused Resources in multi-module setups
First, make sure that when you run lint on a specific
module instead of the full project, we only delete
unused resources in that specific module, and second,
make sure that we don't treat resources in a module
as unused if they are only referenced from other modules.
Finally, make sure that we don't delete resources
referenced from tests.

Test: Unit tests included
Fixes: 141578652
Change-Id: Ia36e9932953f7a86e7ca59a8b38d62468d744188

GitOrigin-RevId: 163e4232f245206720bc2f69d5a17b53e5935899