[Dsl,.toml] Turn VersionCatalog Toml parsing flag on
This turns on read-only Dsl model support for version catalogs
expressed as Toml files, in the limited case where that is a single
catalog defined in a file named libs.versions.toml in the project-wide
gradle/ directory.  This will allow tools such as the Project
Structure Dialog to find and display dependencies defined in version
catalogs and referred to in build files, but new dependencies added
using the PSD will be written in build files directly, and attempt to
change dependency versions of dependencies defined in Toml will fail
silently.

Bug: 200280395
Test: existing
Change-Id: I3c5f6728f0a30ee5f51f4e76208530e7026b6335

GitOrigin-RevId: c5eece70daa742957ad1431c39725bf4bd713a53