[sync] Throw exceptions for too-new AGPs
Catch the case where users attempt to run too-new versions of AGP for
their current Studio, and make sure that that produces a reasonable
build issue (and fails sync).

This must be slightly different from the too-old and
incompatible-preview cases, where we can attempt to upgrade the
project; the Upgrade Assistant is fundamentally incapable of
downgrading projects, so instead simply offer a link to a
compatibility table and rely on the user to sort out their project for
themselves.

Bug: 217593057
Test: added and adapted unit tests, manual
Change-Id: I308acdc6a87e8b29effde6504b999281c7f680a0

GitOrigin-RevId: a7fb0d41ef17d58c1e4b976147ace5fb904ce756