Reduce some boilerplates in `TreeDetailsView`
Not everything has to be classes and inheritance. This change refactors
comparators to be out of functions and compositions, getting rid of the
boilerplate fields and getters. The comparators are stateless, so
unparameterized one (e.g. on name) can be static values.

In addition, there was an intermediate Renderer class created just for
a helper mapping from `FilterType` to `SimpleTextAttrute`. This change
replaces that with just a function. Dispatching on enum should be
straightforward and faster than map lookup, and if this code is ever
converted to Kotlin, the `default` case should be provably redundant.

Bug: n/a
Test: existing
Change-Id: I738c905f74f29a4f03c0917567f6545b0e951f97

GitOrigin-RevId: 9ed36cb512f63e906117eb38d423c7c2ad6653d4