Make screen density consistent across AVD manager
This change fixes an inconsistency between the device edit screen and
the choose a device screen in AVD manager. The problem is users see one
type of screen density on one screen, but then a different density in
the edit screen when they edit the device.

Each device starts with a pre-defined screen density, which can be
recalculated based on user customization of screen width, height,
and diagonal. The calculation is unfortunately unstable, in the sense
that it won't always output a density matching the predefined density,
even with no changes in the screen dimensions. Therefore, the
calculation should only be done when users actually customize the screen.

This change makes it so AVD manager remembers the pre-defined density
and uses it for display purposes unless one of the screen dimensions is
customized by the user, only then does it perform the density
calculation.

Test: manual
Fixes: 143676761
Change-Id: I642c825228f9a050027945c8e1ab165b14b2666d

GitOrigin-RevId: fd2f4157fad52ca938dd487d0b12f101c10f8ad9