Add support for REPLACE as a function call
REPLACE is converted to a token by the lexer,
and then it cannot be used as an identifier even
though the context is enough to disambiguate.

This change does two things:

It uses token remapping to modify the type
of the token at parsing time, moving it from
REPLACE to IDENTIFIER

Re-highlights identifiers at a later stage.
Because the standard highlighter uses only
lexer information, an Annotator is added
to re-visit the identifiers later.

Fixes: 189129263
Test: Included

Change-Id: Ief3e3d3b5ad402c3952f7a5e9506aa5cd1ba4adc

GitOrigin-RevId: 1b62341e25490e8ee90771b03d9df931c60b9bc5