Add light sync reference test project
The light sync test framework (`AndroidProjectRule.withAndroidModels`)
has not been updated to produce MPSS like projects and, therefore,
produces test projects that cannot be created in production.

With upcoming changes it will be updated to create MPSS like projects
and we need to make sure such divergence does not happen in the future.

This change adds a test project, which will be used a reference project
for tests verifying light sync results.

Bug: n/a
Test: n/a
Change-Id: I8010314e4e7afb3e5934221cb941a45509dcfc48

GitOrigin-RevId: 2f55622a1b3f6f2ecb11f9c93208dd64c9fa144a