Unignore PairedDevicesPanelTest
The call to setApplication blocked the deletion of the AVD manager so
PairedDevicesPanelTest was ignored

One thing led to another and I decided to redo PairedDevicesPanel and
the test to be more inline with the rest of the device manager. A big
bonus is that the UI isn't rebuilt with every change in the paired
devices.

Bug: 223246111
Test: PairedDevicesPanelTest
Change-Id: I81a54aeb35aacca34273722feee208b44a9e6d3b

GitOrigin-RevId: d682bc545204239424cb0814c5359d118ee9c2f3