Do not run additional native sync for V2 native models
When build an app with native modules for deployment we make sure that
models for the target device ABI are present and perform additional
native only sync if needed. However, this is not required for native
V2 models. These models are written to files during build and do not
require syncing.

Bug: 217256205
Test: n/a (manually)
Change-Id: Ib504a66433046c75202f6507e4ea13fc1f844574

GitOrigin-RevId: 546d1f57e4590299cbd2592fbaadfff70248e789