[Dsl,.toml] Adjust a test, add a test
When making edits to existing version catalog entries, we will
probably mostly want to be using the result model rather than the
(resolved) property model directly.  This already works for version
references, so add a test for that (and modify the current test for a
non-reference to use the result model too).

(We are currently unable to convert between reference and
non-reference forms because of the difference in external property
name.)

Bug: 200280395
Test: added, existing test adjusted
Change-Id: I40b5321659550e550551c6d6dc14d6d828bf4084

GitOrigin-RevId: 3ccff4075018bb10e9895e874c5c2bb4fc51969b