Rework light sync tets framework for MPSS mode
Light sync test framework simulates what Gradle phase of sync does and
re-uses the IDE phases to set up a test project. This change updates the
simulation to match the MPSS mode and adds a test that compares the
result of real Gradle sync with the result of the light sync, which is
expected to produce the same structure.

The light sync test framework does not support setting up Kotlin facets
yet and so the test project does not apply the Kotlin Gradle plugin.

Also, any test code that fails because of this change is updated.

Bug: n/a
Test: LightSyncReferenceTest
Change-Id: If48087f2318abb143677c4a650642ba099d617fd

GitOrigin-RevId: 6aa23d5c08c6d1a7e307674f708eabc3d3289518