[Compose] Fix insertion of Modifiers.
We had a logic of inserting "Modifier" string if users only types an extension function. In a new version Kotlin plugin does this itself, however it's not working for insertion with tab, actually working only for '\n' insert symbol. Keep our old logic for tab case, because it's annoying that you have this logic on enter and not for tab.

Fixes: 215699716
Test: ComposeModifierCompletionContributorTest
Change-Id: I7d137d7ad7a536520e21650ca6e12bdd2db65cfa

GitOrigin-RevId: 300c32c22074c24c26182070f4d437ce377273ec