[PreviewPicker] Add inspection check for id:<device_id>
When a device is defined by an ID in a @Preview declaration, verifies
that the ID corresponds to an actual Device in the Sdk installation.

This change will cause the inspection to highlight and issue when the ID
is not recognized and will suggest to change it to the default ID used
in the Preview.

Moved CheckResult class to a separate file.

Refactored repeated setup in test file.

Bug: 220006785
Test: PreviewAnnotationCheckTest
Change-Id: Idf67886c325fb1e99feb473e2efbbd78918efd35

GitOrigin-RevId: 10941d8423745fccff23630bda3f4bd93579d997