Add view id information to Visual Linting error messages
When a view is mentioned in a Visual Linting error, if that view has
an id, this adds the id next to the view name. That will make it
easier to identify which view is concerned by the error when there
are many views of the same type in a layout.

Bug: 227251759
Test: Visual Linting tests updated
Change-Id: I10bf80128884ff477f88007f6f4b662056c9382a

GitOrigin-RevId: e321b3f0fb4a4c94dc829120e91993dc74547ebd