Fix execution of Preview refresh to ensure once at a time
The behavior of collectLatest in the refresh flow was supposed to
cancel a running refresh Job when creating new ones, but it was
actually cancelling the wait on the Job, but not the Job itself.

This change fixes this by always selecting the latest RefreshRequest
and skipping the previous requests that were not yet processed. And
also making sure to wait for the running request before collecting
another one.

Fixes: 225058589
Test: manually by following the steps described in the bug.
Change-Id: Ic0252f066188e621864a865e5cc73ca81fba2991

GitOrigin-RevId: 69fbdb7c843ab891fcadb9c0cffc4e3606eb951a