[Dsl,.toml] Support setting parts of Toml maps/literals
In this highly multilingual world, we need to be more careful of the
context of generating literals: being certain either to set the value
we are given, or to create an expression in the correct context (which
might not be the holder element of the model).

This is sufficient to support setting the versions of all three forms
of plugin notation.

Bug: 200280395
Bug: 227793240
Test: existing test adjusted
Change-Id: I7d0b707ba2dd690c6354b38ea085b28e5caa51de

GitOrigin-RevId: 4565a697e4e523000350ebc463016f98462f82a3