Propagate artifacts for module dependencies
and split `IdeModuleDependency` into three different interfaces
representing different states/phases in/of the module dependency
resolution process.

(1) `IdeModuleDependency` represents a resolved module dependency which
    holds exact data needed to address a source set module in the IDE,
    i.e. a build id, a Gradle project path and a source set name.

(2) `IdePreResolvedModuleLibrary` represents a dependency which the AGP
    resolved into a dependency onto another Android project and thus can
    resolve the target IDE source set.

(3) `IdeUnresolvedModuleDependency`represents a dependency on a non-
    Android artifact which is referred to by its file path.

In this change we do not yet implement module dependency resolution and
thus a newly extracted `resolveModuleDependencies` function simply
maps any source sets other than `testFixtures` to `main`, which is
consistent with the existing behavior.

Bug: 228152556
Test: update snapshots
Change-Id: If7868791a4b3e02c9add348bd45f6f908cea2403

GitOrigin-RevId: f8476c9c725bc08b8b9d95cc6a157e2cc35b2747