Fix bug when switching device in device dropdown menu
When switching back to previous device from custom or avd device, the
device dropdown menu doesnt change. For example:
1. Select "Phone"
2. Select "Custom" or any avd device.
3. Select back to "Phone"
The select device will still be "Custom" or the avd device.

This bug happens because setting device to custom or avd devices
doesn't change the specific device. Use setDevice() instead of
SetEffectiveDevice() to solve the issue. setDevice() function use
the default state when argument "prevState" is set to false, which is
what we expected.

Fixes: 214011868
Test: manaully (UI-related)
Change-Id: I9036c0756ccb5897afd137685ff68658149b8467

GitOrigin-RevId: 3aef969a5c5cd69b8a33714c45514120bde48260