[.toml] Parse errors from version catalog syntax errors
At the moment, the error message does not uniquely identify the file
in which the syntax error appears.  If we find a file in the gradle/
directory relative to the project root whose name is the name of the
version catalog with the suffix ".versions.toml", then we assume that
that is the file containing the error; otherwise, we collect the error
message but do not make it navigatable.

Bug: 184090375
Bug: 226959801
Test: added
Change-Id: Id68f963fc3ae5c21ebcb63b5b91f34dbb072ece8

GitOrigin-RevId: 9f48b8e68cbec4a2702a3e69bfba73803e721707