[AUA] Fix UI string bug
The groupingName of an AGP Component Refactoring Processor needs to be
overrideable by localization at runtime, not just at
class-instantiation time.  This implies that we should override it
with a field with a get() method, rather than a bare value.

(This is a somewhat theoretical bug at the moment as I know of no
attempt to localize these UI strings.)

Bug: none filed
Test: none
Change-Id: I519ba11e9bb5b7f6f4e77c0d89c8064ba2049e98

GitOrigin-RevId: 21a1a85c02f8564391856d491f3eee06018b862a