Adapt PreviewAnnotationInspection to MultiPreview
Some of the existing inspections that are executed for Preview
annotations should also be executed for MultiPreview, and some others
shouldn't. Also, MultiPreview added a new possible target for the
Preview annotation (annotation classes), that should be considered
when running the inspections.

This change modifies the existing inspections to consider what
was mentioned above.

Fixes: 228383127
Test: InspectionsTest.kt
Change-Id: I79152674fcf86a1ae4c53001da88e26fbd171678

GitOrigin-RevId: c7deac22003be08b3a9054ba7080b12ee7ebe024