Ctrl-Click To Add/Remove Filter
Ctrl-Click on Tag/AppName/Level to add or remove a filter.

This feature is behind a StudioFlag

Note that adding & removing a filter to an existing filter properly is not trivial. For example, if the existing filter contains
logical operators & parens, just appending/deleting the filter term can result in unexpected results or even in an invalid filter.
For example: If the existing filter is "package:mine | level:error", trying to remove "level:error" will result in an invalid filter

"package:mine |"

Since this is an extreme use case, we don't attempt to be 100% correct when add/removing filters. We just verify that the resulting
filter is a valid filter. If it's not, we disable the action.

Bug: 191880898
Fixes: 229266360
Test: Added
Change-Id: Id1cde10495e5ea6b3c7b69e234951239b86fa543

GitOrigin-RevId: 2954609619a7343af143dd814a27e937692bd2e7