[Surface Launchin] Change type warnings reporting
If the complication description in AndroidManifest.xml contins any unrecognised type, it will be reported as a warning.
The user will be able to choose and launch any other type.
All warnings are reported after errors.
If there are any whitespace characters in the SUPPORTED_TYPES field of
Android Manifest they will be removed before parsing the types.

Tested: Unit test.
Bug: 229723771
Change-Id: Id4bf69ea18caf853ab9fe67c803c6d6ad24da3f2

GitOrigin-RevId: fd54800cbef5718a2c509071ddda45dde216af94