Implement delegate for cached properties dependent on states
In the codebase for capture details, there's a common problem where
there's a property that's expensive to compute that depends on some
state (internal or external). The code has been dealing with this by
manually modyfing the property at the right time to be in sync with
the state.

This change implements an alternative to handle this problem. We
lazily compute and cache the property, and only recompute the property
if the state is different since last time. This is a trade off: it's
less error-prone, and it reduces computation if we keep changing the
state but never query the property, but there's some overhead in
remembering the state.

Bug: n/a
Test: add new
Change-Id: I2e7055b2e75dc4e69091cd5efb5035a596ce095c

GitOrigin-RevId: 866dc68075723a129342d28b0d76d6c3668eb2e1