Fix regression on Compose ViewModel error handling
In Change I3e3fb09f, we decided to report ClassNotFoundException on
inflate calls instead of logging them as errors that can be handled by
our issues model. However, we ended up not logging all exceptions, which
caused a regression in handling view model errors in Compose Preview.

This change fixes this behavior by checking the exception type before
deciding if we should log it or show it to the user.

Fixes: 229093005
Test: N/A. Manually verified ViewModel errors are handled gracefully
Change-Id: I5eaea1ad322e005102bdee37e346ad51609fc917

GitOrigin-RevId: 7f6726db14d318de545178c621f3a0e5a579101c