[Sync] Allow disabling activation of Build Output window on failure
Tools within Studio which explicitly manage the execution and
reporting of sync might wish to manage the display of sync failures to
the user in other ways than the default of popping up the Build Output
Window.  This change allows those tools to opt out of the pop up.

(The Build Output window will continue to be popped up when syncing a
newly-created project, a behaviour for which there is no current hook
in the platform.)

Bug: 227793240
Test: none
Change-Id: Iff7369638700fe7198270df90c71dfa37048897d

GitOrigin-RevId: 49c49c5291980a605483913755e095f87a309e0b