Per-component memory usage reporting implementation.
Memory usage report collection is performed with two passes over the heap. On the first pass we enumerate objects in topological order. On the second pass we determine the belonging of objects to components and update the per-component reports.

Traversal algorithm is implemented in HeapSnapshotTraverse.java
Components are listed in ComponentsSet.java

Test: HeapAnalyzerTest.java tests heap traversal with different types of references
Bug: 230260857
Change-Id: Ic65bf01bc6eb76e613aa705bcea078edd57066b4

GitOrigin-RevId: 45401c4359d15c1c40f6be9d4cfd6ddfeeced282