[AUA] Collect enabled state of UI Controls into enum
We have a version combo box, various buttons to activate the
refactorings (via preview and not), and a refresh button.  The refresh
button is conceptually controllable by whether the tool is in the
loading state or not (it should always be enabled if the tool is not
loading, and never if it is loading), but the other two controls are
more variable depending on the exact state of the tool.  However, it
will never be correct, at least in the current design, for the combo
to be disabled while the run buttons are enabled; we should never be
in a state where the user is prevented from modifying the version in
the dropdown while still being given the option to run the
refactorings.

Bug: 227793240
Test: existing (no change in functionality)
Change-Id: Ib09b172b86465de526fae92cc4b28046a259338f

GitOrigin-RevId: d0662da777b98926e3a6763f4e508a1935f6d5c8