Always apply lint annotation model to design surface
We always want to update lint annotation model to the design surface,
but sometimes Intellij platform may not call
LintHighlightingPass.applyInformationToEditor(). This results missing
issues in the issue panel.

Use UIUtil.invokeLater() directly after collecting the annotation
informations to make sure the annotation is applied to design surface.

Also ignore the progress indicator. "LintHighlightingPass" does not
really do the highlighting -- we use it to update the lint annotation
model for DesignSurface. It should never be canceled once it starts.

Bug: 227392967
Bug: 227389512
Test: manually
Change-Id: Ife3424bca26d1fafd615783945f362c856d8173e

GitOrigin-RevId: 165f5a5927c9a5ec97dab80af0cbb0594149b0ec