230626781: SuspiciousIndentation errors stick around
Make the suspicious indentation check only run from batch
mode, or for unmodified (and not recently edited) files
in the IDE editor. There's an option to make it always
run.

This works around the problem that this check often shows
up right around the caret where you're actively editing,
and particularly for large files with slow analysis, the
markers can stick around even while the code is valid the
analysis finally re-runs.

(This is the IDE-side of the change which implements the
is-edited check for the lint client used by the suspicious
indentation check over in lint.)

Test: Unit tests included
Fixes: 230626781
Change-Id: If22d76d20a261a3022af05c74755d73999e45b3b

GitOrigin-RevId: 2361cd4ce79969d51bf4575709ec96a712b524de