[AUA] Do not execute blocked processors
In general, the UI to these processors should prevent the attempt to
run the processors at all (which is done in the next CL), but in
general if we do get to the point that we are trying to execute a
blocked refactoring, we should not perform the probably-breaking
changes.

The behaviour on preview (which, again, users should not see unless
there is a UI failure) is to allow the usage view.  A subsequent
attempt to run the processor will be ineffective, though, even if the
user deletes all usages that are problematic (in general, deleting
usages is insufficient).  A subsequent CL removes the run button
on processors with blocked components from this view.

Bug: 191813691
Bug: 231690925
Test: none
Change-Id: Ie0a3c8d3b2938dda8b03636ba20400770ea4630a

GitOrigin-RevId: 3d7334a010c4f4263ac04c32d653c605dfc2b170