Stop saving both global/thread times in `CpuTreeNode`
We always update `CpuTreeNode` with clockType, and only want the
numbers for that clockType afterwards. In places (e.g. BottomUpNode),
when we update for one clockType, the other clockType's numbers are
garbage. This change simplifies this, and `CpuTreeNode` only holds the
times for the most recently updated clockType and range.

This change shifts the garbage numbers to the FlameChart's `node`.
This is not a problem, because the numbers were previously also
garbage for one clockType that was not needed. A future change
to FlameChart will fix this problem.

Bug: n/a
Test: update existing
Change-Id: Ieb40659426837342b321987b299c7de022648e9e

GitOrigin-RevId: 3e5abab6ea1e18452dff047cd2596e36ee6e19ce