Do not show the suppress file in issue panel and validation tool
For a visual lint issue, we show all affected files in the issue
description panel and highlight them in the validation tool when
selecting the issue. If all variant files suppress that issue, then
it is not shown in the issue panel.

However, it is possible that only part of variant files suppress the
issue. In that case we should not show the suppressed files in the
"Affected files" section of issue description. We should not highlight
the previews of suppressed files when selected the issue neither.

Bug: 230617078
Test: manually (UI-related)
Change-Id: I11f60a1c71a57766afd1595de43ce8f3bbb41cbf

GitOrigin-RevId: 0594d381a3dd083ee33cf12fe3b772f5d1c4ea41