First version of background visual linting
This version of the background linting runs for the layout file
currently on display, and whenever the Layout Validation tool window is
closed.
It runs all visual lint checks for the reference devices category.
The current interaction with the Layout Validation is simply to hide the
results of background linting when opening Layout Validation, and re-run
background linting when closing Layout Validation.

Bug: 232544714
Test: to be added
Change-Id: I616b0faf9b166512ed3c5ddc41fadfbcea0349ee

GitOrigin-RevId: 5aef9698115ad821e491061050d78262d5d00e02