Refactor and move StringResourceRepository.
This class underlies the StringResourceTableModel class which is an important part of some actions, (e.g. RemoveKeysAction).

In order to make classes like this Action more testable, we must add some abstraction to the StringResourceTableModel and by extension, the StringResourceRepository so that we can build an in-memory fake for testing purposes instead of relying on bringing up the entire IDE and making changes to a real file system.

Bug: 232444069
Test: Added unit tests for StringResourceRepository which were absent before
Change-Id: I3099b1aa31b6f8d0762103d14bdb759104e058c2

GitOrigin-RevId: 909992d90202acb8f869b48698003bda56c30c5b