Remove unnecessary getKey() method from StringResource.
StringResource should not need to know what its own key is - it should exist independently
of how it is keyed in various other classes. Removing other classes' reliance on getKeys is the
first step to removing the key itself from StringResource.

This is one part of a larger refactor that may see StringResourceData removed altogether and
folded in with StringResourceRepository as these two classes serve a very similar purpose and
their existence together makes things harder to understand.

Also, remove the Project reference from the StringResource as it does not need to know which
project it belongs to - it already has that via the StringResourceData reference it has.

Bug: 232444069
Test: Existing unit tests
Change-Id: Ie27e2b86286bca5f51988b1b9792874b7e0532f1

GitOrigin-RevId: 92b7958a91a59f42876c8dabab6dd655ce8fda7f