Inline StringPsiUtils.get{Default,}StringResourceFile.
This provides such limited functionality beyond what
StringResourceWriter already provides that it just makes sense
to use StringResourceWriter. Note that this is an intermediate
change and thus some of the resulting code will not last long,
as such it has not been optimally written (e.g. multiple
returns for the same value).

This will be followed by a change to fold the remaining
methods of StringPsiUtils into StringResourceWriter.

Bug: 232444069
Test: Existing tests
Change-Id: I3fd6d0d1117440ce6fd9b8eda3f6e811eb9672a6

GitOrigin-RevId: 98c2ee97437b7f78226a8ce1843dd04e1c10f9ef