[Dsl,.toml] Actually execute tests
The test cases in the previous case were good; it's a shame that some
of them weren't run.  Actually run them, and fix the problem that they
reveal with multiple entries in an inline table.

The fix for multiple entries in an inline table, to add commas, is a
bit painful in two ways.  Firstly, that we need to explicitly add the
commas ourselves (and so we have to acquire somehow a valid PsiElement
denoting the right kind of comma); secondly, that we need to detect
for ourselves when we shouldn't add one (when we have just inserted
the first element into a TomlInlineTable).  The solution in this CL is
not necessarily the best possible solution, merely one that works.

Bug: 184090375
Bug: 200280395
Test: existing tests run, added
Change-Id: I0337dd1fa10755a8fa72ac0ab029941e90bf2e10

GitOrigin-RevId: 1e63c461318eff182701bf03cc4a559f49f7c1e4