[AUA] Implement a post-upgrade report
We need to behave differently when performing the first sync
subsequent to changing the build files from the Upgrade Assistant.
Define a callback from the Upgrade Assistant refactoring
processor (where it is about to request sync), and use it in the
ContentManager to track that a post-upgrade sync is expected.

Implement new UI states in the Upgrade Assistant ContentManager:
RunningUpgradeSync, which looks exactly the same as RunningSync (but
has a different identity; that may or may not end up being necessary)
and UpgradeSyncSucceeded, which we reach if the sync was both
successful and immediately after build files were upgraded.  Display
information about the upgrade process that just happened in the panel,
as well as suggestions for next steps.

Bug: 217378292
Test: added state transition tests
Change-Id: I1f93b1e9a8fd1efcfdd9b558f0274bdd47210118

GitOrigin-RevId: 504167da78e1d6ed84ef5211e05282c9266d8893