Handle Errors From Logcat Process
If the Logcat process quites with an error (for example https://cs.android.com/android/platform/superproject/+/master:system/logging/logcat/logcat.cpp;l=1068), there will be a pending message still left in the MessageAssembler.

This pending message will contain the last legitimate message terminated by a `\n\n` followed by the error text.

Here we extract the error message if it exists and log it as a system message.

Bug: 191880898
Test: Added
Change-Id: I9d90d1fd371f6d43e59b9196b7283e455594501a

GitOrigin-RevId: 5c6909486bcddc8c1d1fb2b4f178c306dde34c30