Add priority system for render actions
This replaces the standard queue for tasks in RenderExecutor with a
priority queue. It is such that higher priority tasks are executed
first, and if the queue reaches capacity, lower priority tasks are
discarded first.
In the case of equal priority, tasks created first will be executed or
discarded first.

Bug: 232544714
Test: test added
Change-Id: I9f2ba25d4d2ae94a4af1fe6d9fd7718c1b43d8e5

GitOrigin-RevId: 45c2ef4865cf17ebdf4fcfcf511b697cd5b5d903