[AUA] Add a state transition test for internal errors
Refactoring Processors generically check for write permission, so we
can trigger the CaughtException state by turning off write permission
to a file that would be altered.

Actually testing this requires wrapping the code branch used in unit
tests that runs the processor in the same try/catch that is used in
production, which is probably a good thing all round.

Bug: 217378292
Test: added in this CL
Change-Id: Ie14d90ac09fc1b9709292b743d80733988c7a803

GitOrigin-RevId: dc5b7baf02cedb6a1b05f3857574644b7f817de0