Add support for multiple devices
Enable ForegroundProcessDetection to handle multiple devices.
Each device corresponds to a stream.

Each time a device is connected we register an event listener
for LAYOUT_INSPECTOR_FOREGROUND_PROCESS events from that device.

When a device becomes the selected device, we send a
START_TRACKING_FOREGROUND_PROCESS event to it. If there was a
selected device before, we also send a
STOP_TRACKING_FOREGROUND_PROCESS to that device. This makes sure
that the on-device polling happens only on the selected device.

The SelectedDeviceModel introduced in this change will become
more useful later on when for the UI we will need to know at
any time which device we are listening events from.

Test: ForegroundProcessDetectionTest
Bug: 232094819
Change-Id: I73483f19b33c67b52ca6bf8cf5ec05f2cd5cf085

GitOrigin-RevId: 7152b9c2993410409393a926a542673f48c556bb