Do not force-refresh file system
when looking for a project to which a build issue belongs. The root of
an open project must be present in the FVS without refreshing it.
However, getting a directory info may trigger a read action and needs to
be wrapped in such in case in runs in a background thread (it usually
does).

Bug: n/a
Test: n/a
Change-Id: If87d0c41cb0443f7d4a9309bfc039848ac360a22

GitOrigin-RevId: 0d7158ace47e2ed42481d0230650c2f2c5fed41d