Refactor filters for simplicity and to remove dependence on Presentation.
Previously, filters needed to understand how a Presentation works.
Instead of having them know how to update a presentation, provide
methods to get an icon and a description from the filter so the
UI can call these to update itself.

This change also fixes a bug in which the text filter could match
the column holding whether the resource is translatable or not.

Also add unit tests for each of the filters.

Bug: 232444069
Fixes: 234752591
Test: Unit tests for each class
Change-Id: I113d0056933de93aa6cf6ff5871f7731154806a1

GitOrigin-RevId: 7453e9dee851a94407c1fc1256ede25eeb1d280c