Create a ADBLIB Version of ScreenRecorderAction
A lot of this code is a pure copy from ddm code.

The main differences are in:
* ScreenRecorderAction
* ShellCommandRecordingProvider
* EmulatorConsoleRecordingProvider

This functionality depends on a few components in the `intellij.android.core` module, so it cannot be in the `intellij.android.adb` module next to `ScreenshotAction` where it belongs so for now, we'll put it in the Logcat module.

 The dependencies are:
 * `EmulatorAdvFeatures.emulatorSupportsScreenRecording()`
 * `AndroidSdks.tryToChooseSdkHandler()` (used by the above)

Regarding tests, I only ported the existing tests. Coverage was very limited.

Bug: 191880898
Bug: 227213508
Test: Limited coverage only
Change-Id: I063192982abe98f617d26d1db051fab034f8b4c4

GitOrigin-RevId: 114bd54f29ece86477701d39dad3dc8952d1e6fc