Android: remove `studio-analytics-proto` from platform libs
This library is only needed in android plugins. It is included into Android plugin itself as `withProjectLibrary` artifact. Design-tools plugin should not include this library. In Android Studio this library becomes available to Design plugin as a transitive dependency of the Android plugin. Let it be the same in IDEA.

(cherry picked from commit 2ca89bdaed96d851f156124571d61c11936af8d0)

IJ-CR-25210

GitOrigin-RevId: ee906dbf4b2de8986ee79ddd583cdcad4cf2dd4b