Execute and handle handshake
This change adds logic that initiates the handshake with the on-device
library each time a new device is connected.

If the handshake is successful the device supports foreground process
detection. Only in this case Layout Inspector starts foreground process
detection.

If the handshake is not successful, we fall back to the old
implementation of Layout Inspector, where the user has to hand-pick the
process they want to inspect.

Test: ForegroundProcessDetectionInitializerTest, ForegroundProcessDetectionTest, SelectedDeviceActionTest
Bug: 232067042
Change-Id: I642c2b1e0cc55436cc2609c932dcd9883b8e948c

GitOrigin-RevId: 482010a16e13ed0b39f2007953099c0144c4fd2a