[PreviewPicker] Support Orientation as an independent parameter
Orientation changes are now handled separately.

This means that the orientation is now a property of DeviceConfig that
may be changed without affecting the width/height values, so when
instantiated, the orientation must be explicitly set, otherwise it will
default to Portrait.

Width/height/orientation might get treated differently when serializing
to/from a DeviceSpec String, depending on what format is being used.

To DeviceSpec:
- Legacy format (declares Shape), the width and height may be swapped
  to reflect orientation
- DeviceSpec Language format, the orientation parameter might be
  omitted if it matches the width/height relationship

From DeviceSpec:
- Legacy format, orientation is decided by width/height relationship
- DeviceSpec Language format, from orientation parameter, otherwise by
  width/height relationship

Bug: 200290947
Bug: 220005155
Test: updated
Change-Id: I36ce0de6d789cb1830b87eaf3706685bc001a0d8

GitOrigin-RevId: a95e02370e4603607025df92c6568545806ac7ce