[PreviewPicker] Support 'parent' parameter
After selecting an existing device in the picker, making changes to
orientation, should not change the entire device spec into a custom
spec. Instead it should indicate the id of the original device and the
selected orientation.

This can be accomplished by making use fo the 'parent' parameter in the
DeviceSpec Language.

The goal is to keep a reference to the originally selected Device as
long as the user doesn't modify any parameter that changes the
underlying Device (such as width, height, density, shape).

E.g:
- selecting Pixel 4
  - device = "id:pixel_4"
- selecting landscape oentation
  - device = "spec:parent=pixel_4,orientation=landscape"
- modifying width to 1200px (the underlying device is no longer Pixel 4)
  - device = "spec:width=1200px,height=2280px,dpi=440,orientation=landscape"

Bug: 220006785
Test: DeviceIdParameterRuleTest, updated all other
Change-Id: Ied3eabfe2f5ac4ac1be79d1a2d837aa884694c8a

GitOrigin-RevId: 58a2620f5bbddfad6a0bd37b10f5bf8723750199