Fix no issue panel when the first opened file is compose
The shared issue panel is setup by IssuePanelService. However,
IJ's framework doesn't create the instance of IssuePanelService before
it is used.

To avoid this problem, use a StartupActivity to explicitly get the
instance of IssuePanelSevice.

Layout Editor doesn't have this issue because there is a toggle issue
panel button in Layout Editor, which calles the IssuePanelService to
get the count of issue. The IssuePanelService is used at that moment.

Fixes: 235316289
Bug: 229686370
Test: Add regression test and test manually.
Change-Id: I6fc105dc653af209a28196c2bf95254fd18b46e9

GitOrigin-RevId: 1a0a9a70032e5a774e0cce0d90cfb5bbd2bb8981