Make ShowLogcatListener & ClearLogcatListener Use Serial Number
Use serial number instead of an IDevice so that Logcat V2 can remove the ddmlib dependency.

The usage of serial number is inline with how Logcat V2 works already, but for legacy Logcat, we need to do a little work to find the actual IDevice for a given serial number.

Bug: 191880898
Test: n/a
Change-Id: Ic4ec1fe48b34f4ec66fcf0195ff78dc0d0c69acc

GitOrigin-RevId: 54c936bd881ce151a3c5f60c3e7a0e4168b1a27b