Fix highlights countdown
There was an issue where:
- InspectorModel.decreaseHighlights
could set maxHighlights to zero while:
- InspectorModel.update
was running.

With the unwanted timing of:
1) the `update` method could read a value for `maxHighlight` that
   was not zero and therefore skip the scheduling
2) the `decreaseHighligths` could stop rescheduling and set
   `maxHighlight` to zero.
3) the `update` method could immediately after set `maxHighlight`
   to a high number.

That timing would essentially stop the highligh decrease method to
be scheduled, and result in the highlights be left on forever in
the Layout Inspector image.

Fix this by extending the WriteLock section in `decreaseHighlights`.
Also reset the number after reconnecting.

Fixes: 235383531
Test: Unit test added.
Change-Id: I55aa15ff96233b7cff0bdd927a0013d8ce1c4659

GitOrigin-RevId: 885a2ab1fda17539bb35007d145d1e925b4fcb20