Send addRule command upon user action
This change makes it so that we immediately send an addRule command to
the inspector when user clicks the add button in the UI.

Secondly, changed usages of addRule command to updateRule when it's
appropriate. (they are treated the same by the agent, but we should
still call the right command)

Thirdly, this change enables and disables rules by sending an update
evente, as opposed to adding/removing rules.

In addition, this change introduces logic to differentiate table events,
so we don't send reorder command when the action performed isn't related
to reordering.

Test: RuleDetailsViewTest
Bug: n/a
Change-Id: If46dc7ba08c00e660afd7802bdc532fd008987b9

GitOrigin-RevId: a8758fd788e3d88d5ab4f7e450b44d2ada9d3341