[AUA] Improve CaughtException state slightly
Make the constructor accept a String, so that the caller does not have
to make a StatusMessage object themselves.  Make the corresponding
information in the detailsPanel include a link to the local history
dialog (offering a button at that point is risky because we are in a
very uncertain state, but allowing the user to inspect the state of
their files and how they have changed since the start of the upgrade
is probably a good idea).

Bug: none filed
Test: added
Change-Id: Ic9b98d1a4fbcdda820f3c53cf7182823dbd2048f

GitOrigin-RevId: 9e6e3ee5f1d1b6720f7ca3e8a6333148cd3c0d37