Add inner class section to heap reports
The implicit reference to the instance of the enclosing class is a
frequent source of memory leaks. It is suspicious when objects
are retained solely through these references.

This change adds a section to the heap reports that includes a histogram
of these retained objects as well as a histogram that attempts to
attribute those objects to the specific inner class that retains them.

This could be done precisely by computing dominators separately for each
inner class instance, but this can be cheaply approximated by, for each
object O in the set R of objects that are retained by all of the inner
class parent references, walking up the parent pointers from O until
an object is found that:
  * is an inner class
  * is not in R
  * its child on the path to O is in R

Change-Id: I5448e8a460a80a44497eb7ac4446f7ac8fccc480

GitOrigin-RevId: 57daa66d79e633c95729cb3a2e23b8482e79951e