Prevent race condition
Whenever a notification shows in the "Welcome" window, it will attempt
to close the existing notification panel. However, the test only tries
to open that panel once, so it gets stuck if a notification shows up
after it's been opened. The fix involved the following:

- Prevent the Proguard notification from showing up
- Stop explicitly calling "CheckForUpdate" since the platform already
  does this automatically, and each one produces its own notification

As a result of this, UpdateTest can now be re-enabled, but I want to do
that as part of a separate commit in case a rollback is needed.

Fixes: 234170016
Test: this is the test
Change-Id: I70cb4e823934e5b87c0e12e863b556066b27bdbc

GitOrigin-RevId: 743970e83529126725b098eaa7d715fb83d383f4